requirements.txt
kafka-python
time
json